src/requirements/requirements.txt
django==2.1.7
psycopg2-binary==2.7.7
djangorestframework
djangorestframework-gis
